+++ terumet/init.lua
@@ -313,6 +313,8 @@ end
 terumet.do_lua_file('material/concrete')
 terumet.do_lua_file('material/coalproc')
 
+if terumet.options.vac_oven.VAC_FOOD then terumet.do_lua_file('material/vacfood') end
+
 local INTEROPS = {'armor', 'doors', 'unified_inventory', 'tubelib', 'dungeon_loot', 'moreores'}
 for _,mod in ipairs(INTEROPS) do
     if minetest.global_exists(mod) then terumet.do_lua_file('interop/'..mod) end